requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pymysql
cryptography
requests
pytest
pandas
apscheduler